Stasis Bag and Sleeper Changes
This should add a little depth to these two rather shallow machines. Sure sleepers have some drugs and are the only way to do dialysis, but could probably benefit from being more useful, for being a bigass machine attached to the floor in the middle of medical. And stasis bags are a little silly considering they make the occupant completely immune to all harm from any source indefinitely.

Sleepers: These now actually deserve the name, and have an adjustable Stasis setting that has 0%, 50%, 80%, 90%, 99% settings. This can be read as "ignore 99% of life() ticks on this mob", or "ignore every other life() tick on this mob" (for the 50% setting), which is how the stasis bags function now (except they ignore every life tick). This allows you to buy time to treat someone, or have them process drugs more slowly so you can handle multiple patients. On the highest (99% slowdown) setting, it also slightly reduces time-of-death decay, and will allow the patient to be defibbed for longer. (Ody Sleeper: Has the same '30%' effects as a Stasis Bag, see below. No features removed.).

Stasis Bag: These act like the sleeper with a permanent fixed 30% setting, ignoring 2 out of 3 every life ticks. This is a direct nerf to their functionality, considering they can preserve someine indefinitely for now, without fail, against any harm. To try and make up for this, they now feature a built-in injector system, making them a sort of distant cousin to the sleepers. You can load a syringe with 15u of your favorite drugs and insert it into the cryobag in advance (then just re-fold it back up). Upon being used, the syringe will empty into the patient automatically (and the drugs will process at 30% speed, due to their slowed life() ticks). You can remove the syringe later with a screwdriver if you want to change it. If you're a completely crazy person, you could have a 'brute' stasis bag and a 'burn' stasis bag or something with different drugs.

So now the Sleeper, Ody Sleeper, and Stasis bag are basically tiers of the same concept:
- Stasis bag: Portable, but max 15u drug administration, fixed 30% metabolism rate, one time use (but fits in a pocket)
- Ody sleeper: Portable, advanced drug administration, fixed 30% metabolism rate
- Fixed sleeper: Fixed, advanced drug administration, dialysis, 0-100% metabolism rate, ToD decay delay